All parsers were consolidated under the Parsley.Parsers namespace, formerly known as Parsley.Primitives.
