+++ src/Parsley.Tests/QuantifiedParserTests.cs
@@ -1,4 +1,4 @@
-﻿using Parsley.Primitives;
+﻿using Parsley.Parsers;
 using Shouldly;
 using System.Text;
 using Xunit;
