+++ src/Parsley/Grammar.cs
@@ -1,4 +1,4 @@
-﻿using Parsley.Primitives;
+﻿using Parsley.Parsers;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
