+++ src/Parsley/Parsers/OptionalParser.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Parsley.Primitives
+namespace Parsley.Parsers
 {
     public class OptionalParser<TItem> : IParser<TItem>
     {
